Only set the parameter when the experimentalRSC feature is set
