+++ code/frameworks/nextjs/src/portable-stories.ts
@@ -17,7 +17,6 @@ import type {
 
 import type { Meta, ReactRenderer } from '@storybook/react';
 
-import * as rscAnnotations from '../../../renderers/react/src/entry-preview-rsc';
 // ! ATTENTION: This needs to be a relative import so it gets prebundled. This is to avoid ESM issues in Nextjs + Jest setups
 import { INTERNAL_DEFAULT_PROJECT_ANNOTATIONS as reactAnnotations } from '../../../renderers/react/src/portable-stories';
 import * as nextJsAnnotations from './preview';
@@ -49,8 +48,6 @@ export function setProjectAnnotations(
 // This will not be necessary once we have auto preset loading
 const INTERNAL_DEFAULT_PROJECT_ANNOTATIONS: ProjectAnnotations<ReactRenderer> = composeConfigs([
   reactAnnotations,
-  // we only provide the RSC decorator, but not the feature flag. Should be set by the user
-  { decorators: rscAnnotations.decorators },
   nextJsAnnotations,
 ]);
