--- code/renderers/react/src/entry-preview.ts
@@ -1,4 +1,0 @@
-export const parameters: {} = { renderer: 'react' };
-export { render } from './render';
-export { renderToCanvas } from './renderToCanvas';
-export { mount } from './mount';
